+++ site/src/jsMain/kotlin/com/deepshooter/blogmultiplatform/pages/admin/Create.kt
@@ -417,7 +417,7 @@ fun CreateScreen() {
                                        author = localStorage["username"].toString(),
                                        title = uiState.title,
                                        subtitle = uiState.subtitle,
-                                       date = Date.now().toLong(),
+                                       date = Date.now(),
                                        thumbnail = uiState.thumbnail,
                                        content = uiState.content,
                                        category = uiState.category,
